requirements.txt
python-dotenv==1.0.1
requests==2.31.0
yt-dlp>=2025.10.22
pydub==0.25.1
pygame==2.6.1
audioop-lts==0.2.2
